refactor: fix consistency
