+++ tests/test_unit/test_cdf_tk/test_commands/test_dump_data.py
@@ -21,7 +21,7 @@ from cognite.client.data_classes import (
 
 from cognite_toolkit._cdf_tk.client.testing import monkeypatch_toolkit_client
 from cognite_toolkit._cdf_tk.commands import DownloadCommand
-from cognite_toolkit._cdf_tk.commands.dump_data import AssetFinder, EventFinder, FileMetadataFinder, TimeSeriesFinder
+from cognite_toolkit._cdf_tk.commands.download import AssetFinder, EventFinder, FileMetadataFinder, TimeSeriesFinder
 from cognite_toolkit._cdf_tk.utils.file import read_yaml_file
